AMP-7583
git-svn-id: file:///svn/trunk/amp@11210 9e8d75aa-d0cb-44a9-8341-a4b8ba1a38b4

+++ WEB-INF/src/org/digijava/module/widget/action/ShowSectorByDonorChart.java
@@ -129,7 +129,7 @@ public class ShowSectorByDonorChart extends Action {
         }
 		//WIDTH
 		if (form.getImageWidth() == null){
-        	opt.setWidth(new Integer(420));
+        	opt.setWidth(new Integer(450));
         }else{
         	opt.setWidth(form.getImageWidth());
         }
